face_recognizer/requirements.txt
# Явно указываем версию Protobuf, которая удовлетворяет TensorFlow (<5.0)
# и является зрелой версией 4.x. Это предотвратит случайное обновление.
protobuf==4.25.3

# Пакет для обратной совместимости, который требует retina-face
tf-keras

# --- gRPC и Инструменты для генерации кода ---
# Используем стабильную версию gRPC.
grpcio==1.62.2

# Важно! Используем ту же версию для инструментов, что и для основной библиотеки.
# Эта версия генерирует код, совместимый с protobuf v4.
grpcio-tools==1.62.2

# --- Основная библиотека проекта ---
# Ваша главная зависимость.
deepface==0.0.95

# --- Остальные зависимости ---
pillow==11.3.0
pillow-heif==1.1.0
